service/monitoring-train/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0

# AI/ML 학습용 패키지들
transformers==4.42.4
torch>=2.2.0
accelerate==0.33.0
sentence-transformers==2.5.1
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4

# 파일 처리
python-multipart==0.0.9
PyPDF2==3.0.1
openpyxl==3.1.2
xlrd==2.0.1

# 벡터 검색
faiss-cpu==1.7.4

# 추가 패키지들
datasets>=2.0.0
